factor sub-components out of ImpactDashboard
