+++ src/impact-dashboard/Description.tsx
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+const Description = styled.p`
+  font-family: Poppins;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 11px;
+  line-height: 16px;
+  margin: 10px 0 20px 0;
+`;
+
+export default Description;
