+++ src/impact-dashboard/FormGrid.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+const FormGrid = styled.div`
+  width: 100%;
+`;
+
+export default FormGrid;
